Добавление дочернего презентера задачи
