+++ src/main.js
@@ -1,11 +1,11 @@
 import MoviesModel from './model/movies-model.js';
 import BodyPresenter from './presenter/body-presenter.js';
-import Presenter from './presenter/presenter.js';
+import MainPresenter from './presenter/main-presenter.js';
 
 const bodyPresenter = new BodyPresenter();
 bodyPresenter.init();
 
 const moviesModel = new MoviesModel();
-const presenter = new Presenter();
+const mainPresenter = new MainPresenter(bodyPresenter.container,moviesModel);
 
-presenter.init(bodyPresenter.container,moviesModel);
+mainPresenter.init();
