+++ src/view/movies/movies-container-view.js
@@ -1,4 +1,4 @@
-import AbstactView from '../framework/view/abstract-view.js';
+import AbstactView from '../../framework/view/abstract-view.js';
 
 const createMoviesContainerView = () => ('<div class="films-list__container"></div>');
