+++ src/view/movies/movies-list-message-view.js
@@ -1,4 +1,4 @@
-import AbstactView from '../framework/view/abstract-view.js';
+import AbstactView from '../../framework/view/abstract-view.js';
 
 const createMoviesListMesssageView = (moviesCount) => (
   `${moviesCount ? '<h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>' : '<h2 class="films-list__title">There are no movies in our database</h2>'}`
