+++ src/view/movies/movies-list-view.js
@@ -1,4 +1,4 @@
-import AbstactView from '../framework/view/abstract-view.js';
+import AbstactView from '../../framework/view/abstract-view.js';
 
 const createMoviesListView = () => (
   `<section class="films-list">
