+++ src/view/movies/movies-view.js
@@ -1,4 +1,4 @@
-import AbstactView from '../framework/view/abstract-view.js';
+import AbstactView from '../../framework/view/abstract-view.js';
 
 const createMoviesView = () => ('<section class="films"></section>');
